ENH: Make the last implicit sign conversions explicit
One change is a fix. nifti_read_buffer() passed
(int)(ntot / nim->swapsize) to nifti_swap_Nbytes(), which takes a size_t
count that was widened in 2010 because it "might not fit as int". The
cast reinstated that truncation, leaving most of an image above 2 GiB
unswapped. The cast is gone.

nifti_image_read() took the file size through (size_t), which turns
nifti_get_filesize()'s -1 error return into SIZE_MAX and slips past the
guard below it. A signed temporary is used instead, matching what
nifti2_io.c already does at the same place.

The rest casts an int dimension to size_t at the point of use.

+++ cifti/afni_xml.c
@@ -211,7 +211,7 @@ afni_xml_list axml_read_file(const char * fname, int read_data)
 
       if(xd->verb > 4) fprintf(stderr,"-- XML_Parse # %d\n", pcount);
       pcount++;
-      if( XML_Parse(parser, buf, blen, done) == XML_STATUS_ERROR) {
+      if( XML_Parse(parser, buf, (int)blen, done) == XML_STATUS_ERROR) {
           fprintf(stderr,"** %s at line %u\n",
                   XML_ErrorString(XML_GetErrorCode(parser)),
                   (unsigned int)XML_GetCurrentLineNumber(parser));
@@ -299,7 +299,7 @@ afni_xml_list axml_read_buf(const char * buf_in, int64_t bin_len)
 
         if(xd->verb > 4) fprintf(stderr,"-- XML_Parse # %d\n", pcount);
         pcount++;
-        if( XML_Parse(parser, buf, blen, done) == XML_STATUS_ERROR) {
+        if( XML_Parse(parser, buf, (int)blen, done) == XML_STATUS_ERROR) {
             fprintf(stderr,"** %s at line %u\n",
                     XML_ErrorString(XML_GetErrorCode(parser)),
                     (unsigned int)XML_GetCurrentLineNumber(parser));
